[LIGO-914] Fix negative numbers handling
Problem: passing a value starting with `-` as argument to `ligo`
executable does not work, such value is interpreted as a flag and
execution fails.

This makes a lot of our functionality flawed: inputting
parameter/storage in debugger, compiling contract's storage or
dry-running the contract in LSP.

Solution: it has been proposed to e.g. insert a space to the beginning
of the CLI argument (https://gitlab.com/ligolang/ligo/-/issues/1495),
and it works.

Also I introduced a newtype for CLI arguments, so that we never pass
user input as a CLI argument of `ligo` without protecting it first.