[LIGO-833] Avoid linear lambda meta re-embedding
Problem: currently when `EXEC` produces another lambda, we do not
simply copying the meta from the old lambda to the new lambda, rather
re-embed each single function name from the old lambda to the new
lambda.

This is error-prone (we run `embedFunctionNameIntoLambda` over and over
for the same function names) and takes quadratic time overall.

Solution: simply copy the lambda meta from the old lambda to the new
lambda.