[LIGO-833] Refactor pre+post execution logic in snapshots
Problem: we have some bits of logic in `preExecutedStage` and
`postExecutedStage` that is coupled: on `EXEC` we add some stack frames,
and then remove them.

It is simpler to reason about such logic when it is concentrated in one
place, not scattered among `pre` and `post` places.

Solution: add `surroundExecutionInner` which is like `preExecutedStage`
and `postExecutedStage`, but for the cases when we need both of those.

Pros: increased locality + better separation of concerns.
Plus we can now avoid error-prone operations like removal from stack
frames list, which this commit also does.