[LIGO-714] Reuse `LigoTypeExpression` from `Cli.Json`
Problem: we have our own `LigoType` that we parse from source mapper's
output. However, the json-schema for it is the same as for types from
`ligo info get-scope` one. So, we can reuse some types from `squirrel`

Solution: reuse them

:NOTE: As a part of this ticket I also implemented pretty-printing
LIGO types in the dialect of the current running file