[LIGO-690] Add support for .pligo in Web IDE
Problem: .pligo should be accepted as a PascaLIGO extension in addition
to .ligo, but it is not recognized in the Web IDE.

Solution: Add support for .pligo in the Web IDE. Since the /compile
endpoint passes all file names directly from the compile request to the
LIGO compiler without considering file extensions. So, the only
endnpoint we have to change is /compile-expression, which requires
passing the file extension explicitly to the LIGO compiler.