[LIGO-390] Add `ligo-build` -> `validate-cabal-files` dep
Problems:

1. Currently the entire pipeline is executed even if the cabal file is
not updated. This does not make much sense since build will take the
cabal file, not `package.yaml`, as the source of truth, thus checking
something wrong.

2. That cabal job can push a commit that fixes cabal file. However the
old pipeline does not abort at this point, it is executed to the end,
which may block the new pipeline where the cabal file is fixed, thus
increasing the delay.

Solution: in CI config, make `ligo-build` job depend on
`validate-cabal-files` job. Note that semantically such dependency is
really present according to the problem #1.