[LIGO-768] Implement `validateConfig` in the adapter's part
Problem: we do some `launch.json` validations in the `launch` request.
This is not good because in the `launch` request we can throw only original
DAP errors. However, the situation with malformed `launch.json` file
should be handled right before the `launch` event

Solution: make `validateConfig` handler. We'll call it right after all
config resolvings and throw a `ConfigurationException` if something is
wrong