[LIGO-947] Fix module access in as expression
Problem: When we have something like `x as A.B.t`, our parser fails at
the `.B` and interprets the `A` as a `TypeName`. This happens because
`TypeName` has bigger precedence than a `ModuleName`, so the remaining
`.B.t` becomes a mistake. But by changing this, we interpret the
expression as a `data_projection`, not as the expected
`type_as_annotation`.

Solution: Remove the precedence for the `TypeName`. `type_as_annotation`
is not a `_member_expr`, but rather an `_expr_statement` (see the
upstream parsers), which was probably later changed or added by mistake
there.