[LIGO-647] Pass --ligo-bin-path to ligo daemon
Problem: LIGO spawns LIGO subprocesses in the daemon, but if `ligo` is
not in PATH (or it has a name other than `ligo`), the daemon will act
strangely. For instance, `--version` won't work.

Solution: Pass `_lceClientPath` as an argument to `ligo daemon` using
`--ligo-bin-path`.

I avoided some `liftIO`s in `callLigo` while investigating the problem.