[LIGO-901] Fix hlint check further
Problem: in the current variation, Danger `hlint.rb` check does not
react adequately on lint suggestions, the check fails with a weird
error.

The reason is that previously I tried to make it fail gracefully in case
`hlint` executable is not found (and thus call to it returning non-zero
exit code), however now the case of lint suggestions in effectively
handled in the same way as lint suggestions also cause `hlint` to return
non-zero exit code.

Solution:

First distinguish the case of "hlint is not found" and "hlint
returned suggestions" by passing `--no-exit-code` flag.

Now the handling works as follow:
* "hlint is not found" - produces non-zero exit code, we check for it
  and fail explicitly.
* hlint returned empty suggestions; this is already handled specially
  with `.reject { |s| s == '' }`.
* hlint returned suggestions: this is handled.