[LIGO-647] Spawn LIGO with RIO
Problem: We use the IO monad to first initialize the daemon. This should
be used only for debugging, as it uses `ligo` as the default path (i.e.,
looks in PATH).

Solution: Change `Pool LigoProcess` to `IORef (Maybe (Pool
LigoProcess))` :pepe_sad:, so that we first initialize it with
`Nothing`, and later properly use `RIO`.