[LIGO-928] Add LSP functionality to backend
Problem: The backend needs to be able to function as an LSP server. So
we need to be able to receive and respond to messages from the frontend,
over a Websocket connection.

Solution: Use the websockets library to set up a shim that can receive
messages from the frontend, pass them to the LSP executable, collect the
responses from the executable, and send them back to the frontend.