[#1591] Fix stepping behaviour at end
Problem: if I reach a failing execution state (by executing `failwith`),
the next step causes a prompt with internal error. This happens because
`writeTerminatedEvent` is executed then.

Solution: the reason of the issue is that `PastFinish` runs the same
`writeTerminatedEvent`, we definitely don't want to executed it after
`writeException`.

So let's just terminate the debug session in this case.