[#1592] Avoid debug session termination on reaching end
Problem: we discussed this with @DK318, and actually we both find it
unreasonable that pressing `Next` when standing at the last snapshot
with failure causes the debug session termination.

The reason is: if I step over the contract quickly, spamming `Next` or
`Step In` command, I probably don't want to accidentially exit, I rather
want to step around to find a bug. And letting fall over the edge
of the execution would be an annoying UX.

This is what I wanted to have yet in summer, but in my absence this was
implemented differently in morley-debugger and we copied this behaviour;
now when we revised this, would be nice to do the improvement.

I actually also wanted us to not exit on "ok termination" too, but
currently exiting on it works more or less fine from UX point of view.

Solution: on facing `PastFinish`, if it is after a failure, then do not
cause termination.