[Chore] Fix  support for local modules
Problem:
We can't find a definition of a thing if it's defined in
local module like `x` in
```
module A = struct
  let x = 1
end
```
because in this case `vdef` for `x` will appear in
`mdef` for `A` but not as a standalone declarations in `get_scope` result.

Solution:
Create `unfold_get_scope : get_scope_info -> get_scope_info`
function which recursively traverses local modules and adds all definitions
from this modules to list of declarations.