[#1541] Use `parsePreprocessed` instead of `parse`
Problem: we noticed that we see locations for top-level functions
with some preprocessor directives. After some investigations we've seen
that the problem is wider: we're trying to parse a contract before
preprocessing stage, so, we have some parse errors there.

Solution: use `parsePreprocessed` instead of `parse`. Change types
where it is necessary.