[#1648] Handle parsing errors in the CST
Problem: The current code doesn't deal with with files that have syntax
errors. This is because `~raise` will convert it to `None` in case there
are any errors.

Solution: Make a very permissive raise that ignores everything, since
those errors will be reported via diagnostics anyway. Note that fatal
errors can't be ignored, so the solution was to use a `try...with`
block. One example of such errors is when we can't lex, and sadly, in
such case we will also not obtain a CST.