[#1678] Warn user about LIGO versions
Problem: When LIGO has a version smaller than 0.61.0, we should tell
them to update. Also, the hints when LIGO was not found are not
sufficient.

Solution: Add a hint in case LIGO was not found and tell the user it
must support `ligo lsp`. In case of an empty string, assume the user
knows what they're doing. For an unsupported version, ask the user to
update.