[#1541] Miscellaneous stepping polishing
Motivation: Heitor found some inconvenient points in debugger's stepping
after his testing. This comments fixes some of them (and adds some tests).

Here is the list of what was polished:
* Now we can stop at local function assignments (as statements). Note
that we still need to polish it a bit but it would be resolved as a part
of #1684.
* If `EventFacedStatement` and `EventExpressionPreview` snapshots have
the same locations then we'll record only a snapshot with `EventFacedStatement`
event. The reason is simple: having 2 steps with the same effect is
inconvenient. We can skip `EventExpressionPreview` and show `EventExpressionEvaluated`
immediately.

:NOTE: point with missing statement location for `TestId.One.id` function
from `poly.mligo` contract is still actual. We can't fix it now just
because of missing location from source mapper's output.