[#1658] Support getting configuration from VSCode
Problem: We currently ignore the VSCode configuration defined in
`package.json`.

Solution: Add support for getting the configuration on startup (but not
yet for configuration changes). Initially, we support the maximum number
of problems (diagnostics) as a PoC that this works and in the next
commits, we'll add support for other configurations.