[#1658] Respect logging verbosity
Problem: We have `debug` which controls whether to send debug messages
or not. Let's generalize it and have `logging_verbosity` to control
which messages to filter.

Solution: Replace `debug : bool` with `logging_verbosity :
MessageType.t` and use it to control when to send logging messages to
the user. Also update the VSCode config to support this.