[#1684] Strip duplicate locations from typed contract and rethink location recording
Problem: there're some problems with the current duplicate locations
stripping mechanism. For e.g. we may have locations
`a`, `b`, `a` in dfs upper-bottom order. After current deduplication
we'll have `b`, `a` locations, so, some statements will appear
in the wrong order (like with statements in `switch` or local
functions assignments).

Solution: move stripping to typed contract and delete duplicates
only in groups (e.g. in dfs order we have `a`, `a`, `b`, `a`, `a`, `b`.
After stripping `a`, `b`, `a`, `b`). Note that after deduplication we'll
still have some duplicates, so, I've reimplemented location recording
mechanism in such a way, that it records only the first occurrence of
a location.