[#1658] Support `deprecated` in the configurations
Problem: We have `--deprecated` which is used to enable PascaLIGO, but
we don't really handle it on the server side.

Solution: This isn't the best solution as Emacs and Vim don't officially
support a configuration, so we always enable it for those two editors.
For Visual Studio Code, however, we read the configuration and pass it
to `get_scope`.