[#1533] Add command for using last granularity
Problem: we are at the point when we need quick ways to switch between
granularities. A particular use case: something is wrong and I want to
temporarily switch to `expression` granularity.

The only currently available option - press the button with a mouse and
in the opened pane select the needed granularity (with mouse or by
entering the option number + pressing Enter) is still quite annoying.

Solution:

One convenient addition (aside from others that will be applied in next
commits) is let the user switch to a different granularity in one
action, without opening the pane.

Here there are the options:
* Cycle through the all the available granularities. Which means 1
  action to switch to the desired granularity, and 2 actions to switch
  back (or vice versa).
* Cycle only among two options. For sure we will leave `statement`
  granularity, but whether will the particular user in particular case
  prefer `expression` or `expression (pre + post)` granularity - is a
  question.
* Switch to the last used granularity. This seems to perfectly fit the
  considered use case, and the user will be able to switch between those
  two granularities which they want to use.

So in this commit I start remembering the last used granularity and add
a command for switching to it.