[#1744] Make hovers prettier and using modern format
Problem:
We're sending hover message in `MarkedString`, which is deprecated.
Also, we're not limiting max width of a hover if it's a type info.
Also, it's not always clear type of what object is hovered, because
there is only type in hover.

Solution:
Change functions that print type so they return `PPrint.document`.
Set width limit to 60 when rendering this document for hover.
Add name of the hovered term to message, e.g. `x : int`.
This is also useful for investigating failures of hover tests.