[#1655] LSP: Implement range formatting for toplevel statements
Problem:
Our formatting currently erase comments, so users may want
to run formatting only on some range in a file.
There is a `range formatting` LSP request that allows this.

Solution:
Implement range formatting via replacing all toplevel statements
in given range with theit pretty-printed version.
We should do the same with at least statements in local modules.
To obtain ranges of toplevel statements, some changes in LIGO are
required:

First, in some dialects there is no function that extracts
range of a toplevel declaration in CST module.

Second, `Region.cover` was supposing that given regions are not
intersecting, while in JsLIGO we can insert a semicolon
with region intersecting the rigion of a statement during self-pass.

Also we need to expose a function that prints a declaration for
each dialect.