[#1532] Show `layout comb` in types and fix sum types creation
Problem: we want to use LIGO types for decompiling Michelson values
into LIGO ones. At this moment we don't distinguish types that have
`tree` layout and `comb` one. Moreover, there is a bug with sum type
creation: now it's just like `fromLigoTableField . toPairs` and it's
not correct because it doesn't care about the order.

Solution: support `layout` field in `LigoTypeTable`, update pretty printers
and mind the order in sum types.