[#1532] Strip outdated invalidation calls
Problem: currently if I step through the contract quickly enough,
I will visit the new snapshots earlier than we manage to convert
variables to LIGO format. But we will still invoke `InvalidatedEvent`
once the conversion is done, despite it is already for an old snapshot.

Cons of this is that it imposes an extra overhead - exchanging variables
information between VSCode and the adapter requires quite a lot of
messages.

Pros is that invoking outdated `InvalidatedEvent` still has some effect
- if the current snapshot contains some of the values from the old
snapshots (oftentimes it does), then those values will be updated
earlier, and that's convenient from UX point of view.

Solution: I'm in struggle, but as a try I implement aborting
`InvalidatedEvent` if we made a step.

It's to discuss whether to we want to eventually include this commit or
not.