[#1532] Fix bug with wrong fields ordering in records
Problem: records are parsed from JSONs into hashmaps. It means that
`toPairs/toList` on hashmaps won't save the order of fields.

Solution: write `buildLigoValue` function instead of `Buildable` instance
that takes `LigoType` and `LigoValue`. `LigoTypeField` has a field
`_ltfDeclPos` which stores a position of this specific field, so, we
can use it to restore the order.