[#1532] Make `Copy as Expression` copy LIGO value in specific dialect
Problem: at this moment we display LIGO values in some convenient format
(e.g. map looks like `[a -> b, c -> d]`). Actually, this is not a valid
format for any LIGO dialect. For e.g. in CameLIGO it should look like
`Map.literal([(a, b), (c, d)])`. Motivation here is to copy this value
and paste it in some contract in two clicks.

Solution: support copying values in specific LIGO dialect via
`Copy as Expression` feature in the variables pane.