[#1658] Send support capabilities based on enabled or disabled features
Problem: We should not indicate that we support some feature if that
feature was disabled, or we might get unwanted behavior.

Solution: Use the list of disabled features when taking each capability
configuration into account. We only mention them as supported during
initialization if they are not disabled.

Mention in the README as well that the user should restart the language
server after changing this feature.