[#1732] Recognize constant as a statement
Motivation: if a standalone constant appears last in scope then
we don't treat it as a statement. For e.g.
```ocaml
match some_val with
| Some x -> x + 2
| None -> 42
```
if `some_val` is `None` then `42` wouldn't be treated as a statement.
From stepping perspective it's not convenient because we'll just skip
`match ... with` instantly.