[#1694] Disallow looking for defs from stdlib in rename/definition/type definition
Problem: Stdlib definitions have `""` as their file. This causes a
problem when looking for a definition in VS Code, as it will try to open
such file and obviously not work. For prepare rename, we also want to
disallow dealing with stdlib, as we shouldn't be able to rename things
there.

Solution: Look for the file in these three methods. If it's empty,
return `None`.