[#1701] Fix searching of references in other files
Problem: When we look for references, we do a search to check that the
references come from the same URI for each opened file. This also means
that we won't see references in included/imported files, unless that
other file has been opened as well.

Solution: I now collect everything in a map whose keys represent the
file name and whose elements is a set of references of a particular
definition. Now we can group every reference per file that it's seen.