[#1679] Fix invalid extension error when using Ext
Problem: If some command, such as Compile Expression, requires
CommandRequiredArguments.Ext, it might throw an exception. This is
because the following may be returned as a path:

extension-output-ligolang-publish.ligo-vscode-#1-LIGO Compiler

This is because VS Code may return output channels as text editors.
However, we try to take the "extension" of that and get
.ligo-vscode-#1-LIGO Compiler, which is obviously wrong. Now, we use the
extension of the last known contract path if we don't have a LIGO
contract as a path.