[#1766] Don't do clean up on restart event
Problem: at this moment `restart` button is broken: if you try to press
it then you'll see an error message that states that language server
state is not initialized. This occurs because restart event is just a
terminate one with `restart = true` and on terminate event we're doing
a complete clean up of this state and closing a thread pool.

Solution: don't do anything in `onTerminate` handler if `restart = true`.