[#1772] Show `$it` variable LIGO formatted
Problem: now we support LIGO format for variables but we still show
`$it` variable in the Michelson one. Source mapper provides a `source_type`
field for locations, so, we can use these types to decompile `$it` variable.

Solution: extract a `source_type` field in location metas, decompile
`$it` variable on `EventExpressionEvaluated` event.