[#1806][#1769] Record evaluated event only for topmost location
Problem: at this moment we're recording pair events
(`EventExpressionPreview`/`EventFacedStatement` and `EventExpressionEvaluated`)
for different locations. For example, we have the next locations
`a b a`. In the current model we'll record `EventExpressionPreview` for
the first `a` location and record `EventExpressionEvaluated` for the
second one. It's not correct because these locations are pointing to
the different Michelson instructions but have the same source type.

Solution: introduce boolean flag that indicates whether we recorded a
location for the first time. Record pair event only when this flag is
`True`.