[#1805] Launch LIGO LSP even when an update is available
Problem: The extension doesn't launch if the update LIGO dialogue is
open. This is poor UX.

Solution: Now we don't `await` the updating code to run anymore, but
rather launch it in the background. The caveat is: it's possible
`fs.writeFileSync` will run while LIGO is being executed and fail with
ETXTBSY. This is the case even if we run `client.stop()` with a set
timeout. Unfortunately, I was not able to find a good solution besides
waiting a bit (I set 2 seconds, it seems to work well).