[#1762] Implement dumping CST in the compiler
Problem: in the debugger we still use tree-sitter parsers for LIGO
contracts. It's not good because we need to update them when LIGO
adds/changes something in its syntax. It would be better to get CST
directly from the compiler.

Solution: derive `yojson_of` instances for CameLIGO and JsLIGO syntaxes,
implement a `ligo info dump-cst` command. As a bonus introduce minifying
JSON output (stripping redundant whitespaces).