[Chore] Write logs to a file
Problem: We should have a file where we write all of our logs to.

Solution: Use some code I had stashed and make it write to a log file in
the user's temporary directory. This logs everything, including every
single linol RPC request.

The only downside is that the messages logged from the LSP server will
be shown in a verbose manner. For example, if I logged "test test test"
with warning level, it would show as follows:

```
2023-05-26 21:39:31.936664529Z [DEBUG] jsonrpc2: send json: {"params":{"message":"test test test","type":2},"method":"window/logMessage","jsonrpc":"2.0"}
```

They will still show as expected in the user's output channel, though.