[#1762] Get rid of the `tree-sitter` in the adapter's part
Problem: now we have dumping CST functionality in the compiler.
Let's parse its output and convert it to our AST (`SomeLIGO Info`).
After this we can strip redundant code which was used for the `tree-sitter`.

Solution: translate CST types to Haskell, implement `MessagePack` instances
for them and implement a conversion function. Also modify `Info` type.
Now it stores only the range of the node. All the other info is redundant
for the debugger.