[Chore] Decompiler: emit quantifiers
Problem:
There is no `Ast_unified.type_expression` that will express
the type of `List.map` since there are no quantifiers in
CST type expressions. Normally this type expression is an RHS of
a type decl, and there is a nanopass that eliminates all `T_for_all`.
But we want to show the type for `List.map` in LSP hovers, and we
have only type expression (not declaration) here.
Also we don't want to show extra parens like
`map : (a -> b) -> (a list -> b list)`.

Solution:
Discard quantifiers in decompiler (thiscase  should happen only when
decompiling the standalone expression).
Improve the function that adds parens so if `T_Fun` is a RHS of
another `T_Fun` the parens are not added.