[#1819] Use `checkCompilation` in entrypoint validation
Problem: when we validate an entrypoint we use `compileLigoContractDebug`
function which is really expensive. At this stage we don't need a
parsed `LigoMapper`, so, it could be replaced with smth more lightweight.

Solution: implement `checkCompilation` function which calls
`ligo compile contract -e ep file` and use it instead of `compileLigoContractDebug`.