[#1642] Implement semantic tokens for all dialects
Problem: We want to support colorizing tokens using semantic
highlighting/semantic tokens from the LSP specification.

Solution: Generalize folding functions a bit to take any node rather
than just the initial CST node. Perform some fixes in the CST and
folding functions in order to preserve order or allow better
identification of some nodes. Remove the variant hack. Rename the
`true_const` and `false_const` keywords. Finally, implement the
aforementioned method for the three dialects.

The idea is to iterate on every token and give a label to it. In an
ideal world, just matching on each wrap would be enough, but alas, there
were some obstacles, which are summarized here:
* It's sometimes difficult to know whether something is a function name,
  function argument, or an ordinary variable. So we explicitly match on
  some nodes and propagate an "environment" that allows to override what
  we're currently visiting.
* The iteration order of some things are not in increasing region order.
  We needed to explicitly match on those and force the correct iteration
  order.
* Some tokens, like EOF, semicolons inserted with ASI, etc, are "ghost",
  and should not have tokens associated with them.
* Some tokens lie about their types. For example, `kwd_as` which may be
  a colon.

I implemented both semantic tokens full and semantic tokens region for
now. The first one is just a special case of the second one where the
range is the entire file.

The algorithm itself descends the tree into every token, and collects
the start line, start character, length, token type, and token modifiers
(always empty for now) and pushes them into a mutable int vector. After
all is done, this vector is iterated and the start line and start
characters are turned into relative offsets, i.e., the difference of
lines and characters to the previous character, as per the LSP
specification. If the line delta is greater than 0, then the character
delta is the absolute position. Furthermore, some tokens in LIGO are
allowed to be multiline, such as comments and verbatim strings, and so
they must be first broken down into multiple lines and each line
tokenized. This yields a problem: we don't know the length of each line
only looking at the start and stop positions alone (although we may be
able to figure out something by inspecting the payload string, but I
prefered to avoid it as it's O(n)). For now, I hard coded 256 characters
for such situations and I hope that's reasonable.

There are still some things that need to be fixed: comments in JsLIGO do
not always highlight (I think something is wrong directive ranges may
overlap, and they don't store some ranges like of operators in
expressions, and PascaLIGO comments are never highlighted, since they
are not stored in the CST. I believe fixing those is outside the scope
of this MR.

I renamed `false_const` and `true_const` to `kwd_false` and `kwd_true`,
respectively.

Finally, I changed the JsLIGO CST and Parser to deal with a hack in
variants. We were using `MultiArg`, which didn't quite differentiate
what was the constructor or not apart from looking at its head. This
made implementing semantic tokens for variants difficult, so I split it
into three distinct variant cases and propagated the changes across the
compiler (up to nanopasses).