[#1825] Implement copying error messages
Problem: at this moment user doesn't have an ability to copy-paste
the debugger errors. The only way is to make a screenshot but it's
a not convenient way to do that.

Solution: implement a `Copy` button which tries to copy error contents
via external software (e.g. `xclip` in Linux). If the user doesn't have it
then opens a read-only file and uses `vscode` commands to perform the
copying.