[Chore][#1948] Bump `minimal-supported` and `recommended` versions
Problem: in the newest LIGO version most of the critical problems
with the source mapper are fixed. Now we can bump the version.

Solution: make next adaptations:
1. `decl_pos` field in record/sum types is stripped. Now we should deduce
`comb` or `tree` layout by the tree in the `layout` field.
2. locations for let-expressions now point to function body.
3. fixed a bug with the last value on the stack.
4. now variables have qualified names. But if they come from a different
file then the module name would be `Mangled_{path_to_contract}.foo`.
It's too long, so, it would be good to strip them.