[#1978] Adapt `resolve-config` on the frontend part
Problem: the `resolve-config` command is already implemented in the
compiler and adapted on the backend. Let's adapt it on the frontend's
side.

Solution: add a `ligoConfigPath` field for the `launch.json` config.
Implement a new command which creates a config snippet in `CameLIGO`
and `JsLIGO`. Update `README.md` as well.