[#1741] Resolve FIXME in the VSCode extension
Problem: There is a FIXME related to the project root in the VSCode
extension. In reality, the extension wants us to add a custom method to
the language server to fetch the project root, but now we always assume
it's where ligo.json is, so let's do it.

Solution: Use the path to ligo.json as cwd. The language client became
unused in various places, so remove it.