[#2056] Highlight signatures/interfaces properly
Problem: We forgot to highlight some keywords.

Solution: Add them to the keyword lists. There was a problem with
attributes that I found and fixed along the way, where we would consume
things inside `@[` eargerly until we found a `]`, but we want to do it
in a greedy manner, since if there are other `[]` in the same line
(e.g., a list), we'll continue highlighting as attributes incorrectly.