[#1741] Add tests
Problem: We need some more tests to check that we are using the correct
project roots.

Solution: Use `on_doc` tests to open a file, fetch diagnostics, and ask
for the last project root. Check that we have no diagnostics (there
should be no errors or warnings) and that the project root is as
expected by the test info.