[#2056] Add missing syntax highlighting keyword and fix unbounded type annotations with includes
Problem: `extends` was not recognized as a keyword by the syntax
highlighting. Moreover, if an `include` followed some `val`, it would
get highlighted as a type.

Solution: Add `extends` to the list of JsLIGO keywords. Add `include` to
the end of a type definition/annotation.

Also, `implements` was duplicate in the list of JsLIGO keywords, so I
removed one of them.