[#2056] Handle "go to declaration"
Problem: There is a third confusingly named request: go to declaration.
The specification doesn't say how this request should be handled, but I
think it's fair to make it assume the old behavior of go to definition.

Solution: Declare such request in the server, and implement it by simply
returning the place where the reference came from. This mimics the old
behavior of go to definition. So from a reference, such as `M.x`, we
visit `x` within `M`, while in `I.x`, we visit `x` within `I`, while go
to definition and go to implementation would behave different for each
case.