[#2056] Handle references for signature items
Problem: References for signature items aren't dealt with correctly.
This also means that rename and prepare rename don't work as intended.

Solution: Refactor some functions from `Go_to_definition` that will be
useful for us in `References`. Now, we can find a subgraph of the mod
graph that contains all modules that are reachable from one another
somehow using the WCC algorithm, and for each module in this graph, get
all of its definitions. Now, we can look for the definitions that match
the name and level (type, term, module) of the thing we're trying to
find a reference for. We change the existing functions in `References`
to now deal with finding all references of multiple definitions rather
than just one.

For `Rename` and `Prepare_rename`, we just reuse the functions defined
in `References` to handle the changes.