[#2056] Find implementers from signature names
Problem: If we have a definition like `module type T = ...`, and we look
for implementers of `T`, we won't see useless information.

Solution: Check if we are looking for the implementations of a
signature. If so, just return the name of the module implementing that,
rather than looking at its defs.