[#2056] Simplify resolve_mod_name
Problem: `resolve_mod_name` is polymorphic so it can use `string` or
`Uid.t` as the module path. Previously, we just used the `Uid.t` form,
but then we wanted the `string` one as well. Not only it's unneeded (we
can always use `Uid.t`), but it'll also be useful in the next commit.

Solution: Remove polymorphism and use `Uid.t`; propagate changes.