[#2056] Handle `include`s
Problem: `include`s are not handled for modules and for the top-level.
Moreover, there is a bug with module aliases pass: we are replacing the
resolved names for impls/exts with the ones from the base module.

Solution: For module aliases, we now look into the list of module
aliases for the greatest location smaller than a given location that
matches the same name. This simply means: the last thing in scope
(assuming that the scopes are ordered).

For extends, add a new `extends` field to `mdef` and handle it similarly
to `implements`. Add it to the mod graph, and change go to definition
appropriately. For all scope passes, use `include_mvar` to inline the
child definitions into the parent module.