[#1811] Remember names and structure for named functions
Problem: at this moment function types in JsLIGO are decompiled into
named functions with wildcard names. We want to preserve that info.

Solution: store a list of names inside `Ligo_prim.Arrow.t` and use it
for decompiling.

Also, store a mark inside `curry` nanopass for the first
argument in lambda. It would be helpful for inferring types.