[#1957] Fix go to type definition for inferred records
Problem: Go to type definition doesn't work properly for record types,
it instead points to the record expression body rather than the type
definition.

Solution: In the typer, use the location of the `orig_var` rather than
the one provided by `create_type`, which will point to the body. Note
that if there is no `orig_var`, it will keep the old behavior.

I also added a few tests to ensure it works.