[#2056] Fix finding defs within a module
Problem: When we look for definitions, we find everything that matches a
name and level. In reality, though, we want to find just one thing
matching a name and a level: the one which was defined last.

Solution: Find the maximum definition by the ordered location, rather
than returning all of them.

Add a test ensuring that this regression doesn't return.