[#1650] Implement document symbol
Problem: We want to add the document symbol request to the language
server.

Solution: Add a new request for this symbol, which traverses all
definitions while gathering some useful information to be displayed for
the symbols.

There is a need to group these symbols in a hierarchy, and for that, I
created a rose tree. The idea of the algorithm is as follows: sort def
ranges by their start symbols, and if they are equal, give priority to
the longest one (the one whose end is farther). This has an interesting
(but unproven AFAIK) property that the ranges are now ordered such that
the "parent" is always before the children, and we can take every
following range while it intersects this parent, and use this to create
the rose tree.