[#1650] Replace body_range with decl_range
Problem: We don't really use body_range for anything, and it would be
more useful for us to have the declaration range of things.

Solution: Rename body_range to decl_range and use the range provided
from a declaration or signature item.

There was a small bug in the parser where a signature declaration range
was incorrect, so I fixed it here as well.