[#1650] Nest destructured decls
Problem: When we have a declaration like `let a, b = _1` our hierarchy
will show `a > b > _1`, but we should show `a, b > _2` instead, where
`a` and `b` will be inside `_2`.

Solution: Change the rose tree so that it will have a non-empty list of
elements rather than a single element: those that all compared as equal.
Now, we can change document symbols so that, if rose tree node has one
element, we keep the original behavior, otherwise we consider it as
destructured and provide some more information and better nesting.