[#1650] Protect against invalid ranges
Problem: The request throws an exception if `selectionRange` is not
contained in `range`.

Solution: Use `None` if that happens.

BTW, I simplified the intersection so that we can better handle ranges
that just barely touch at the end, using the knowledge that the
intersection always uses that the first one is less than the second.