[#1647] Implement document highlight request
Problem: we do not yet implement Document highlight request.

Solution: implement this request handling.

The request is about highlighting all occurances of an identifier in
the currently opened document, so I implement this on top of references
finder, adopting it to work in single file.

Yet do not account for highlight kinds, left a todo for that.