[#1676] Add function type parameters
Problem: in expressions like `let f (type a) (x : a) = (x : a)`
we don't have a type variable definition for `a`. This leads to
not working `go to definition`. Moreover, sometimes it may point to
the wrong definition (e.g. when an abstraction name shadows an existing
definition).

Solution: handle this variable in `E_type_abstraction`.