[#2113] Refactor: unite cases of configs receival
Problem: I need to add some logic after the configuration is updated in
LSP. Currently I'd have to write it twice - in `Ok [config]` and `Ok
configs` cases of the pattern match.

Solution: unite those cases.

Actually, LSP specification says that we should get exactly the same
number of configs in that response, which we previously put in the
config request, so always 1, and `Ok configs` case is normally
impossible. But just in case, let's preserve it.