[#1922] Extract `ligoBinaryInfo` to commons
Problem: we have name of the config option for LIGO path mentioned in
many places as magic string constant.

`ligoBinaryInfo` which kinda can be used as a constant that carries this
path, it is lsp-only and debugger has to use its own constant.

Solution: extract `ligoBinaryInfo` to commons and use it everywhere.