[#1922] Do not reuse `config` variable
Problem: when `config` (`WorkspaceConfig`) is updated, the change goes
straight to the persistent storage and somehow the in-memory value is
not updated. I.e.

```
config.update(LIGO_PATH, newVal)
config.get(LIGO_PATH)
```

returns the old value.

This affects the code in `updateLigo.ts`, where we may change the
path to LIGO in settings, and immediately then access that path setting
again.

Solution:

By using `vscode.workspace.getConfiguration()` we can get the actual
config value.

So never carry `config` value across functions in `updateLigo.ts`, pick
config from scratch every time we need it.

Inline `updateLigoImpl` into `updateLigo` as the latter became a direct
call of the former.

Note that we carry the `config` variable a lot in `updateLigo.ts`, but
de-facto it used a couple of places - once to poke the `ligo` and see if
it works, once when changing that path + once to poke `ligo` again. This
shouldn't be too expensive. Moreover, there are other places in the code
where we just call `vscode.workspace.getConfiguration()` when we need
a portion of config.