[#1922] Suggest restarting VSCode on `ligo` path change
Problem: when path to `ligo` in settings is changed, the LSP extension
does not handle the new path automatically.

Solution: track when path to `ligo` changes, prompt a suggestion to
restart VSCode when it happens.

Possible alternative solutions:

At first, I wanted to restart only the necessary parts of the LSP without
restarting the entire extension or VSCode.

However, really most of the LSP needs to be re-initialized in either
way: two `LanguageClient`s (w/ and w/o semantic tokens) that contain
path to LIGO, commands should be removed and set again as they refer to
`LanguageClient`, buttons must be at least temporarily hidden as they
refer to commands.

Another option would be to manually call `deactivate()` and
`activate()`, i.e. restart the entire extension from code.

On this matter, I'm now understanding one small reason why, on extension
re-install, VSCode requires full VSCode reload - freeing all the
resources properly is not so easy, it requires saving all the produced
`Disposable`s. Currently we ignore pretty lot of `Disposable`s, and in
`deactivate` we do not wait for the `stop()` calls. I suppose VSCode
authors didn't want to rely on typescript programmer's ability to
cleanup resources properly, and we better don't try to do this ourselves
too.

So I went with entire VSCode restart, similarly to what happens when the
extension is re-installed.
But if we dare to restart extension from inside the extension, this
should be implementable.