[#2083] Print TZIP-16 check warnings in compile storage
Problem: when `Compile storage` button is used, if the storage is not
TZIP-16 compliant we would not indicate that in any way; while
`ligo compile storage` would produce warnings.

Solution: fetch stderr of `ligo compile storage` and other commands and
print it to logs along with the stdout (which is the execution result).

This required switching from `execFileSync` to `execFile`, as the former
does not allow fetching stderr. In a discussion we agreed that this
should be fine and there no strict reasons to prefer `execFileSync`.