[#1807] Avoid auto-filling LIGO path config
Problem: we have a logic that, if it detects LIGO path setting to be not
set, it calls `which ligo` and fills the setting with its result.

There are several issues with this.

1. This feels like a bad UX; we better never touch the config without
asking the user first. It seems better to just understand that if LIGO
path is not set, we should pass `ligo` as LIGO path and let OS resolve
that.

Maybe previously there were reasons for resolving paths, but we cannot
remember them now; if one day we understand that we have to supply full
paths, better still not touch the config and call `which ligo` every
time we see no LIGO path in config.

2. That filling logic is invoked when `if (binaryPath)` is not visited,
where `binaryPath` is what we got from the config.

Perhaps previously this worked, but today we have a default `ligo` value
registered for that option, which means that `binaryPath` is never
`undefined`. The only condition when `if (binaryPath)` will be visitied
is thus when `binaryPath === ""`, which is probably not what we want.

Solution: remove the logic responsible for config changing.

Rely on that LIGO path setting always resolves to something.

We assume that `getBinaryPath` returns an unchecked path (binary there
can be wrong) which seems good for one set of callers and potentially
undesired for others (some commands now may try to invoke the bad
binary). Probably we can live with it.